feat(condo): DOMA-10042 moved suggestServiceProviders to resident domain
