+++ apps/condo/domains/organization/constants/common.js
@@ -427,8 +427,7 @@ const DEFAULT_ROLES = {
     },
 }
 
-const RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC = 24 * 60 * 60
-const MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC = 100
+
 
 module.exports = {
     TIN_LENGTH,
@@ -448,7 +447,4 @@ module.exports = {
     MIN_ROLE_NAME_LENGTH,
     MAX_ROLE_NAME_LENGTH,
     MAX_ROLE_DESCRIPTION_LENGTH,
-    MAX_ROLE_COUNT,
-    RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC,
-    MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC,
 }
