+++ apps/condo/domains/organization/utils/serverSchema/index.js
@@ -13,7 +13,6 @@ const { OrganizationLink: OrganizationLinkGQL } = require('@condo/domains/organi
 const { OrganizationEmployeeSpecialization: OrganizationEmployeeSpecializationGQL } = require('@condo/domains/organization/gql')
 const { RESET_ORGANIZATION_MUTATION } = require('@condo/domains/organization/gql')
 const { REPLACE_ORGANIZATION_EMPLOYEE_ROLE_MUTATION } = require('@condo/domains/organization/gql')
-const { SUGGEST_PROVIDER_BY_TIN_QUERY } = require('@condo/domains/organization/gql')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const Organization = generateServerUtils(OrganizationGQL)
@@ -48,17 +47,6 @@ async function replaceOrganizationEmployeeRole (context, data) {
     })
 }
 
-async function suggestProviderByTin (context, data) {
-    if (!context) throw new Error('no context')
-    if (!data) throw new Error('no data')
-    return await execGqlWithoutAccess(context, {
-        query: SUGGEST_PROVIDER_BY_TIN_QUERY,
-        variables: { data },
-        errorMessage: '[error] Unable to suggestProviderByTin',
-        dataPath: 'result',
-    })
-}
-
 /* AUTOGENERATE MARKER <CONST> */
 
 module.exports = {
@@ -69,6 +57,5 @@ module.exports = {
     OrganizationEmployeeSpecialization,
     resetOrganization,
     replaceOrganizationEmployeeRole,
-    suggestProviderByTin,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
