+++ apps/condo/domains/resident/constants/constants.js
@@ -168,6 +168,9 @@ const MAX_RESIDENT_DISCOVER_CONSUMERS_BY_WINDOW_SEC = 10
 const RESIDENT_FIND_ORGANIZATIONS_WINDOW_SEC = 24 * 60 * 60 // seconds
 const MAX_RESIDENT_FIND_ORGANIZATIONS_BY_WINDOW_SEC = 100
 
+const RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC = 24 * 60 * 60 // seconds
+const MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC = 100
+
 const REDIS_LAST_DATE_KEY = 'LAST_SEND_BILLING_RECEIPT_NOTIFICATION_CREATED_AT'
 
 module.exports = {
@@ -177,4 +180,6 @@ module.exports = {
     RESIDENT_FIND_ORGANIZATIONS_WINDOW_SEC,
     MAX_RESIDENT_FIND_ORGANIZATIONS_BY_WINDOW_SEC,
     REDIS_LAST_DATE_KEY,
+    RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC,
+    MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC,
 }
