+++ apps/condo/domains/resident/schema/SuggestServiceProviderByTinService.js
@@ -2,16 +2,17 @@
  * Generated by `createservice organization.SuggestServiceProviderByTinService --type queries`
  */
 
-const { pick, uniqBy, sortBy } = require('lodash')
+const { pick, uniqBy } = require('lodash')
 
 const { GQLCustomSchema } = require('@open-condo/keystone/schema')
 
 const { CONTEXT_FINISHED_STATUS } = require('@condo/domains/acquiring/constants/context')
 const { AcquiringIntegrationContext } = require('@condo/domains/acquiring/utils/serverSchema')
 const { MeterResourceOwner } = require('@condo/domains/meter/utils/serverSchema')
-const access = require('@condo/domains/organization/access/SuggestServiceProviderByTinService')
-const { SERVICE_PROVIDER_TYPE, MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC, RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC } = require('@condo/domains/organization/constants/common')
+const { SERVICE_PROVIDER_TYPE } = require('@condo/domains/organization/constants/common')
 const { Organization } = require('@condo/domains/organization/utils/serverSchema')
+const access = require('@condo/domains/resident/access/SuggestServiceProviderByTinService')
+const { MAX_RESIDENT_SUGGEST_SERVICE_PROVIDER_BY_WINDOW_SEC, RESIDENT_SUGGEST_SERVICE_PROVIDER_WINDOW_SEC } = require('@condo/domains/resident/constants')
 const { RESIDENT } = require('@condo/domains/user/constants/common')
 const { RedisGuard } = require('@condo/domains/user/utils/serverSchema/guards')
