+++ apps/condo/domains/resident/utils/serverSchema/index.js
@@ -10,6 +10,7 @@ const { generateServerUtils, execGqlWithoutAccess } = require('@open-condo/codeg
 const { getLogger } = require('@open-condo/keystone/logging')
 
 const { REGISTER_RESIDENT_MUTATION } = require('@condo/domains/property/gql')
+const { SUGGEST_PROVIDER_BY_TIN_QUERY } = require('@condo/domains/resident/gql')
 const { Resident: ResidentGQL, REGISTER_RESIDENT_INVOICE_MUTATION } = require('@condo/domains/resident/gql')
 const { ServiceConsumer: ServiceConsumerGQL } = require('@condo/domains/resident/gql')
 const { REGISTER_CONSUMER_SERVICE_MUTATION } = require('@condo/domains/resident/gql')
@@ -124,6 +125,17 @@ async function findOrganizationsForAddress (context, data) {
     })
 }
 
+async function suggestProviderByTin (context, data) {
+    if (!context) throw new Error('no context')
+    if (!data) throw new Error('no data')
+    return await execGqlWithoutAccess(context, {
+        query: SUGGEST_PROVIDER_BY_TIN_QUERY,
+        variables: { data },
+        errorMessage: '[error] Unable to suggestProviderByTin',
+        dataPath: 'result',
+    })
+}
+
 /* AUTOGENERATE MARKER <CONST> */
 
 module.exports = {
@@ -136,5 +148,6 @@ module.exports = {
     getResidentExistenceByPhoneAndAddress,
     registerResidentInvoice,
     findOrganizationsForAddress,
+    suggestProviderByTin,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
