+++ apps/condo/domains/resident/utils/testSchema/index.js
@@ -30,6 +30,7 @@ const {
     DISCOVER_SERVICE_CONSUMERS_MUTATION,
     SEND_MESSAGE_TO_RESIDENT_SCOPES_MUTATION,
     REGISTER_SERVICE_CONSUMER_MUTATION,
+    SUGGEST_PROVIDER_BY_TIN_QUERY
 } = require('@condo/domains/resident/gql')
 const { makeClientWithResidentUser } = require(
     '@condo/domains/user/utils/testSchema')
@@ -257,6 +258,13 @@ async function registerResidentInvoiceByTestClient (client, resident, invoiceRow
     return [data.result, attrs]
 }
 
+async function suggestProviderByTinServiceByTestClient(client, attrs = {}) {
+    if (!client) throw new Error('no client')
+    const { data, errors } = await client.query(SUGGEST_PROVIDER_BY_TIN_QUERY, { data: attrs })
+    throwIfError(data, errors)
+    return [data.result, attrs]
+}
+
 module.exports = {
     Resident, createTestResident, updateTestResident,
     registerResidentByTestClient, makeClientWithResident,
@@ -269,5 +277,6 @@ module.exports = {
     registerResidentServiceConsumersByTestClient,
     registerResidentInvoiceByTestClient,
     findOrganizationsForAddressByTestClient,
+    suggestProviderByTinServiceByTestClient
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
