+++ apps/condo/schema.graphql
@@ -76978,15 +76978,6 @@ type ReplaceOrganizationEmployeeRoleOutput {
   status: String!
 }
 
-input SuggestProviderByTinInput {
-  tin: String!
-}
-
-type SuggestProviderByTinOutput {
-  tin: String!
-  name: String!
-}
-
 input ExportPropertiesToExcelInput {
   where: PropertyWhereInput!
   sortBy: [SortPropertiesBy!]
@@ -77690,6 +77681,15 @@ type FindOrganizationForAddressOutput {
   account: FindOrganizationForAddressAccountType
   hasMeters: Boolean!
   hasBillingData: Boolean!
+}
+
+input SuggestProviderByTinInput {
+  tin: String!
+}
+
+type SuggestProviderByTinOutput {
+  tin: String!
+  name: String!
 }
 
 input SyncTourStepsInput {
@@ -82188,7 +82188,6 @@ type Query {
   }`
   """
   checkUserExistence(data: CheckUserExistenceInput!): CheckUserExistenceOutput
-  suggestProviderByTin(data: SuggestProviderByTinInput!): [SuggestProviderByTinOutput]
   exportPropertiesToExcel(data: ExportPropertiesToExcelInput!): ExportPropertiesToExcelOutput
   allResidentBillingReceipts(where: ResidentBillingReceiptWhereInput, first: Int, skip: Int, sortBy: [SortResidentBillingReceiptsBy!]): [ResidentBillingReceiptOutput]
 
@@ -82270,6 +82269,7 @@ type Query {
   predictTicketClassification(data: PredictTicketClassificationInput!): TicketClassifier
   getResidentExistenceByPhoneAndAddress(data: GetResidentExistenceByPhoneAndAddressInput!): GetResidentExistenceByPhoneAndAddressOutput
   findOrganizationsForAddress(data: FindOrganizationsForAddressInput!): [FindOrganizationForAddressOutput]
+  suggestProviderByTin(data: SuggestProviderByTinInput!): [SuggestProviderByTinOutput]
   exportMeterReadings(data: ExportMeterReadingsInput!): ExportMeterReadingsOutput
 
   """
