+++ apps/condo/schema.ts
@@ -66052,7 +66052,6 @@ export type Query = {
    * }`
    */
   checkUserExistence?: Maybe<CheckUserExistenceOutput>;
-  suggestProviderByTin?: Maybe<Array<Maybe<SuggestProviderByTinOutput>>>;
   exportPropertiesToExcel?: Maybe<ExportPropertiesToExcelOutput>;
   allResidentBillingReceipts?: Maybe<Array<Maybe<ResidentBillingReceiptOutput>>>;
   /**
@@ -66132,6 +66131,7 @@ export type Query = {
   predictTicketClassification?: Maybe<TicketClassifier>;
   getResidentExistenceByPhoneAndAddress?: Maybe<GetResidentExistenceByPhoneAndAddressOutput>;
   findOrganizationsForAddress?: Maybe<Array<Maybe<FindOrganizationForAddressOutput>>>;
+  suggestProviderByTin?: Maybe<Array<Maybe<SuggestProviderByTinOutput>>>;
   exportMeterReadings?: Maybe<ExportMeterReadingsOutput>;
   /**
    * Export of property meters readings. A link to the file is returned
@@ -72788,11 +72788,6 @@ export type QueryCheckUserExistenceArgs = {
 };
 
 
-export type QuerySuggestProviderByTinArgs = {
-  data: SuggestProviderByTinInput;
-};
-
-
 export type QueryExportPropertiesToExcelArgs = {
   data: ExportPropertiesToExcelInput;
 };
@@ -72836,6 +72831,11 @@ export type QueryGetResidentExistenceByPhoneAndAddressArgs = {
 
 export type QueryFindOrganizationsForAddressArgs = {
   data: FindOrganizationsForAddressInput;
+};
+
+
+export type QuerySuggestProviderByTinArgs = {
+  data: SuggestProviderByTinInput;
 };
